Refactor commands, and simplify rate limiter
