+++ src-tauri/src/commands.rs
@@ -25,7 +25,6 @@ use crate::app_in_memory_config::{AppInMemoryConfig, ExchangeMiner, DEFAULT_EXCH
 use crate::auto_launcher::AutoLauncher;
 use crate::binaries::{Binaries, BinaryResolver};
 use crate::configs::config_core::{AirdropTokens, ConfigCore, ConfigCoreContent};
-use crate::configs::config_mcp::{ConfigMcp, ConfigMcpContent};
 use crate::configs::config_mining::{
     ConfigMining, ConfigMiningContent, MiningModeType, PauseOnBatteryModeState,
 };
@@ -40,7 +39,6 @@ use crate::events::ConnectionStatusPayload;
 use crate::events_emitter::EventsEmitter;
 use crate::events_manager::EventsManager;
 use crate::internal_wallet::{mnemonic_to_tari_cipher_seed, InternalWallet, PaperWalletConfig};
-use crate::mcp::audit::AuditLog;
 use crate::mining::cpu::manager::CpuManager;
 use crate::mining::gpu::manager::GpuManager;
 use crate::mining::pools::cpu_pool_manager::CpuPoolManager;
@@ -75,7 +73,7 @@ use std::fs::{read_dir, remove_dir_all, remove_file, File};
 use std::str::FromStr;
 use std::sync::atomic::Ordering;
 use std::thread::sleep;
-use std::time::{Duration, Instant, SystemTime};
+use std::time::{Duration, Instant};
 use tari_common::configuration::Network;
 use tari_common_types::seeds::mnemonic::{Mnemonic, MnemonicLanguage};
 use tari_common_types::seeds::mnemonic_wordlists::MNEMONIC_ENGLISH_WORDS;
@@ -2155,127 +2153,4 @@ pub async fn set_custom_node_directory(path: String) -> Result<(), InvokeError> 
     }
 
     Ok(())
-}
-
-#[tauri::command]
-pub async fn get_mcp_config() -> Result<serde_json::Value, String> {
-    let content = ConfigMcp::content().await;
-    let mut value = serde_json::to_value(&content).map_err(|e| e.to_string())?;
-    if let Some(obj) = value.as_object_mut() {
-        if obj.get("bearer_token").and_then(|v| v.as_str()).is_some() {
-            obj.insert(
-                "bearer_token_redacted".to_string(),
-                serde_json::Value::String(content.redacted_token().unwrap_or_default()),
-            );
-            obj.remove("bearer_token");
-        }
-    }
-    Ok(value)
-}
-
-#[tauri::command]
-pub async fn get_mcp_token() -> Result<Option<String>, String> {
-    let content = ConfigMcp::content().await;
-    Ok(content.bearer_token().clone())
-}
-
-#[tauri::command]
-pub async fn set_mcp_enabled(enabled: bool) -> Result<(), String> {
-    if enabled && ConfigMcp::content().await.bearer_token().is_none() {
-        let token = ConfigMcpContent::generate_token();
-        let now = SystemTime::now();
-        let expiry_days = *ConfigMcp::content().await.token_expiry_days();
-        let expiry = now + std::time::Duration::from_secs(u64::from(expiry_days) * 86400);
-        ConfigMcp::update_field(ConfigMcpContent::set_bearer_token, Some(token))
-            .await
-            .map_err(|e| e.to_string())?;
-        ConfigMcp::update_field(ConfigMcpContent::set_token_created_at, Some(now))
-            .await
-            .map_err(|e| e.to_string())?;
-        ConfigMcp::update_field(ConfigMcpContent::set_token_expires_at, Some(expiry))
-            .await
-            .map_err(|e| e.to_string())?;
-    }
-    ConfigMcp::update_field(ConfigMcpContent::set_enabled, enabled)
-        .await
-        .map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn refresh_mcp_token_expiry() -> Result<(), String> {
-    if ConfigMcp::content().await.bearer_token().is_some() {
-        let now = SystemTime::now();
-        let expiry_days = *ConfigMcp::content().await.token_expiry_days();
-        let expiry = now + std::time::Duration::from_secs(u64::from(expiry_days) * 86400);
-        ConfigMcp::update_field(ConfigMcpContent::set_token_expires_at, Some(expiry))
-            .await
-            .map_err(|e| e.to_string())?;
-    }
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn revoke_mcp_token() -> Result<(), String> {
-    ConfigMcp::update_field(ConfigMcpContent::set_bearer_token, None::<String>)
-        .await
-        .map_err(|e| e.to_string())?;
-    ConfigMcp::update_field(ConfigMcpContent::set_token_created_at, None::<SystemTime>)
-        .await
-        .map_err(|e| e.to_string())?;
-    ConfigMcp::update_field(ConfigMcpContent::set_token_expires_at, None::<SystemTime>)
-        .await
-        .map_err(|e| e.to_string())?;
-    ConfigMcp::update_field(ConfigMcpContent::set_enabled, false)
-        .await
-        .map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn set_mcp_port(port: u16) -> Result<(), String> {
-    ConfigMcp::update_field(ConfigMcpContent::set_port, port)
-        .await
-        .map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn set_mcp_max_transaction_amount(amount: Option<u64>) -> Result<(), String> {
-    ConfigMcp::update_field(ConfigMcpContent::set_max_transaction_amount, amount)
-        .await
-        .map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn set_mcp_tier_enabled(tier: String, enabled: bool) -> Result<(), String> {
-    match tier.to_lowercase().as_str() {
-        "read" => {
-            ConfigMcp::update_field(ConfigMcpContent::set_read_tier_enabled, enabled)
-                .await
-                .map_err(|e| e.to_string())?;
-        }
-        "control" => {
-            ConfigMcp::update_field(ConfigMcpContent::set_control_tier_enabled, enabled)
-                .await
-                .map_err(|e| e.to_string())?;
-        }
-        _ => return Err(format!("Unknown tier: {tier}")),
-    }
-    Ok(())
-}
-
-#[tauri::command]
-pub async fn get_mcp_audit_log(count: usize) -> Result<Vec<serde_json::Value>, String> {
-    let entries = AuditLog::get_recent(count).await;
-    entries
-        .into_iter()
-        .map(|e| serde_json::to_value(e).map_err(|e| e.to_string()))
-        .collect()
-}
-
-#[tauri::command]
-pub async fn export_mcp_audit_log() -> Result<String, String> {
-    AuditLog::export().await.map_err(|e| e.to_string())
 }
